+++ test/module/shared_model/builders/protobuf/test_builder.hpp
@@ -22,7 +22,7 @@
 #include "builders/protobuf/builder_templates/proposal_template.hpp"
 #include "builders/protobuf/builder_templates/query_template.hpp"
 #include "builders/protobuf/builder_templates/transaction_template.hpp"
-#include "module/shared_model/validators/validators_mocks.hpp"
+#include "module/shared_model/validators/validators.hpp"
 
 using namespace shared_model::proto;
 using namespace shared_model::validation;
